Added some code to help sort the forecast errors
Twice now, a forecast has been attributed to me incorrectly. While I can not currently pinpoint the cause, I have added code that adds the username of the forecaster to the start of their forecast's comments. Code has also been added to remove it when displaying these comments; this should make my life easier when it comes to moving them around in the database. Further steps should be isolating the cause.