Update to ForecastProblemLineChart
Wrapped any use of dates inside a new Date(x).toString().slice(0, x) to ensure that, when presenting the data in the chart, they are all converted to GMT. Added an info button to inform users of the fact that all charts are presented using GMT / all forecasts are converted.

Also updated the info button next to the problem name above the Line Chart to include the same information regardless of whether the problem is a single or multi certainty one. It also informs users of the GMT conversion.