Trying a fix for the % of time at this prediction issue
Issue: When a user had submitted at least one forecast before the deadline and at least one forecast after the deadline (happens when someone submits a forecast before I close the problem early), the backend, when calculating the % of time at this score, would be using the submission time of the next forecast along (even if it was submitted after the deadline). This meant that it was giving % longer than possible. What it should do is check if the next forecast was submitted before the deadline, and if it was not, compare to the closeDate instead. This is attempt 1 at fixing